pyproject.toml
[project]
name = "autopics"
version = "3.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow-avif-plugin>=1.5.5",
    "pillow-jxl-plugin>=1.3.7",
    "pyside6>=6.10.2",
]

[dependency-groups]
dev = [
    "auto-py-to-exe>=2.49.0",
]
